Product done

+++ Admin/src/main/java/com/example/admin/Entity/RevenueEntity.java
@@ -19,5 +19,5 @@ public class RevenueEntity {
 
     private LocalDate date;
 
-    private double totalRevenue;
+    private Long totalRevenue;
 }
